server/requirements.txt
appdirs==1.4.4
click==7.1.2
distlib==0.3.1
Faker==8.1.2
filelock==3.0.12
Flask==1.1.2
hazelcast-python-client==4.1
itsdangerous==1.1.0
Jinja2==2.11.3
MarkupSafe==1.1.1
python-dateutil==2.8.1
six==1.16.0
text-unidecode==1.3
virtualenv==20.4.6
Werkzeug==1.0.1
